Lesson 10. Added forgotPassword functionality; Added asynchronous video processing and email sending.

+++ src/main/java/com/phobod/study/vcp/service/impl/UserServiceImpl.java
@@ -1,22 +1,20 @@
 package com.phobod.study.vcp.service.impl;
 
-import java.nio.file.Path;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
-import com.phobod.study.vcp.component.UploadVideoTempStorage;
 import com.phobod.study.vcp.domain.User;
 import com.phobod.study.vcp.domain.Video;
 import com.phobod.study.vcp.form.VideoUploadForm;
 import com.phobod.study.vcp.repository.search.VideoSearchRepository;
 import com.phobod.study.vcp.repository.storage.VideoRepository;
-import com.phobod.study.vcp.service.ImageService;
-import com.phobod.study.vcp.service.ThumbnailService;
+import com.phobod.study.vcp.security.SecurityUtils;
 import com.phobod.study.vcp.service.UserService;
-import com.phobod.study.vcp.service.VideoService;
+import com.phobod.study.vcp.service.VideoProcessorService;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -25,19 +23,11 @@ public class UserServiceImpl implements UserService {
 	private VideoRepository videoRepository;
 
 	@Autowired
-	private VideoService videoService;
-
-	@Autowired
-	private ImageService imageService;
-
-	@Autowired
-	private ThumbnailService thumbnailService;
-
-	@Autowired
-	private UploadVideoTempStorage uploadVideoTempStorage;
-
-	@Autowired
 	private VideoSearchRepository videoSearchRepository;
+	
+	@Autowired
+	@Qualifier("simpleVideoProcessorService")
+	private VideoProcessorService videoProcessorService;
 
 	@Override
 	public Page<Video> listVideosByUser(Pageable pageable, String userId) {
@@ -51,26 +41,35 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public void uploadVideo(User currentUser, VideoUploadForm form) {
-		Path tempUploadedVideoPath = uploadVideoTempStorage.getTempUploadedVideoPath();
-		String videoUrl = videoService.saveVideo(tempUploadedVideoPath);
-		byte[] thumbnailImageData = thumbnailService.createThumbnail(tempUploadedVideoPath);
-		String thumbnailImageUrl = imageService.saveImageData(thumbnailImageData);
-		Video video = new Video(form.getTitle(), form.getDescription(), thumbnailImageUrl, videoUrl, currentUser);
-		videoRepository.save(video);
-		videoSearchRepository.save(video);
+		Video video = videoProcessorService.processVideo(form);
+		video.setOwner(currentUser);
+		video.setTitle(form.getTitle());
+		video.setDescription(form.getDescription());
+		saveVideoToRepositories(video);
 	}
 
 	@Override
 	public void updateVideo(String videoId, VideoUploadForm form) {
 		Video video = videoRepository.findOne(videoId);
+		if (!video.getOwner().equals(SecurityUtils.getCurrentUser())) {
+			throw new AccessDeniedException(String.format("User %s denied access to edit video %s. This user is not the owner of the video.", SecurityUtils.getCurrentUser().getLogin(), video.getId()));
+		}
 		video.setTitle(form.getTitle());
 		video.setDescription(form.getDescription());
+		saveVideoToRepositories(video);
+	}
+
+	private void saveVideoToRepositories(Video video) {
 		videoRepository.save(video);	
 		videoSearchRepository.save(video);
 	}
 
 	@Override
 	public void deleteVideo(String videoId) {
+		Video video = videoRepository.findOne(videoId);
+		if (!video.getOwner().equals(SecurityUtils.getCurrentUser())) {
+			throw new AccessDeniedException(String.format("User %s denied access to delete video %s. This user is not the owner of the video.", SecurityUtils.getCurrentUser().getLogin(), video.getId()));
+		}
 		videoRepository.delete(videoId);
 		videoSearchRepository.delete(videoId);
 	}
